The last line in the REPL is half cut off
This appears to be due to the interplay of the padding on
the #xterm element and the box-sizing of the #terminal-container
element being set to border-box. This issue in the xterm.js repository
talks about the same thing: xtermjs/xterm.js#1283

The workaround mentioned in the issue appears to solve the problem here.
That is, set the box-sizing on the #terminal-container to be content-box.